Initial ugly port to Apache 2.4
The minimum changes required for the modules to compile and work with
Apache 2.4 and pass basic testing.  This breaks compilation on Apache
2.2, which will be restored in a subsequent commit with additional
Autoconf probing.

This duplicates a bunch of code in mod_authldap and removes old code
that will no longer compile with #if 0.  This too will be cleaned up
in a subsequent release.

Reviewed-on: https://gerrit.stanford.edu/292
Tested-by: Russ Allbery <rra@stanford.edu>

+++ modules/ldap/mod_webauthldap.h
@@ -153,7 +153,7 @@ typedef struct {
     MWAL_DCONF *dconf;
 
     apr_table_t **entries;  /* retrieved ldap entries */
-    int numEntries;
+    size_t numEntries;
 
     apr_table_t *envvars;    /* which attributes to place into environment */
     apr_table_t *privgroups; /* which privgroups to check and place into
